Split the manipulator subsystem into the manipulator (elevator and wrist) and gripper subsystems.
