+++ src/main/java/frc/robot/commands/ManipulatorPresetFactory.java
@@ -2,11 +2,13 @@ package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
+import frc.robot.subsystems.gripper.Gripper;
 import frc.robot.subsystems.manipulator.Manipulator;
 import frc.robot.utils.LoggedTunableNumber;
 
 public class ManipulatorPresetFactory {
   Manipulator manipulator;
+  Gripper gripper;
 
   LoggedTunableNumber elevatorRetracted = new LoggedTunableNumber("Elevator Retracted", 0.0);
   LoggedTunableNumber wristRetracted = new LoggedTunableNumber("Wrist Retracted", 0.0);
@@ -27,8 +29,9 @@ public class ManipulatorPresetFactory {
   LoggedTunableNumber elevatorProcessor = new LoggedTunableNumber("Elevator Processor", 1.0);
   LoggedTunableNumber wristProcessor = new LoggedTunableNumber("Wrist Processor", 1.0);
 
-  public ManipulatorPresetFactory(Manipulator manipulator_) {
+  public ManipulatorPresetFactory(Manipulator manipulator_, Gripper gripper_) {
     manipulator = manipulator_;
+    gripper = gripper_;
   }
 
   public Command retracted() {
@@ -103,16 +106,16 @@ public class ManipulatorPresetFactory {
               manipulator.setWristReference(wristIntake.get());
 
               if (manipulator.isInPosition()) {
-                manipulator.setGripper(-3500, -3500, -3500);
+                gripper.setGripper(-3500, -3500, -3500);
               } else {
-                manipulator.setGripper(0, 0, 0);
+                gripper.setGripper(0, 0, 0);
               }
             },
-            () -> manipulator.setGripper(0, 0, 0),
-            manipulator)
+            () -> gripper.setGripper(0, 0, 0),
+            manipulator, gripper)
         .until(
             () ->
-                manipulator
+                gripper
                     .getDistanceReading()
                     .filter(
                         (Double d) -> {
@@ -121,12 +124,12 @@ public class ManipulatorPresetFactory {
                     .isEmpty())
         .andThen(
             Commands.runEnd(
-                    () -> manipulator.setGripper(1000, 1000, 1000),
-                    () -> manipulator.setGripper(0, 0, 0),
-                    manipulator)
+                    () -> gripper.setGripper(1000, 1000, 1000),
+                    () -> gripper.setGripper(0, 0, 0),
+                    gripper)
                 .until(
                     () ->
-                        manipulator
+                        gripper
                             .getDistanceReading()
                             .filter(
                                 (Double d) -> {
